rename Device.py as device_property.py
