+++ dptb/negf/device_property.py
